requirements.txt
# VoiceTrace Dependencies
# Pinned versions for reproducibility
# Generated: 2026-04-29

# Core
anthropic==0.76.0
requests==2.32.5
python-dotenv==1.2.1
pyyaml==6.0.3

# Data processing
pandas==2.3.3
numpy==2.3.5
scipy==1.16.3

# NLP evaluation
jiwer==4.0.0
bert-score==0.3.13
sacrebleu==2.6.0
scikit-learn==1.8.0

# Geocoding
geopy==2.4.1

# Testing
pytest==9.0.2

# Dashboard (optional)
streamlit==1.52.2

# Paper figures
matplotlib>=3.8.0
